requirements.txt
django==3.0
xhtml2pdf 0.2.4
djangorestframework
markdown
django-filter       